Refactor scalog clauses
