+++ core/src/main/scala/io/github/scalaquest/core/parsing/scalog/CodeGen.scala
@@ -1,0 +1,8 @@
+package io.github.scalaquest.core.parsing.scalog
+
+/** A Prolog unit of information. */
+trait CodeGen {
+
+  /** String representation of the information. */
+  def generate: String
+}
